+++ src/main/java/com/myapp/team/TeamApplication.java
@@ -2,8 +2,9 @@ package com.myapp.team;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
 public class TeamApplication {
 
     public static void main(String[] args) {
