+++ src/main/java/com/myapp/team/user/config/CustomSecurityConfig.java
@@ -1,28 +1,23 @@
 package com.myapp.team.user.config;
 
 import com.myapp.team.user.service.CustomUserDetailsService;
-import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 
 @Configuration
 @EnableWebSecurity
 public class CustomSecurityConfig {
+
 
     @Bean
     public UserDetailsService userDetailsService() {
@@ -35,14 +30,14 @@ public class CustomSecurityConfig {
                 .csrf(AbstractHttpConfigurer::disable)
                 .authorizeHttpRequests((authz) -> authz
                         .requestMatchers("/admin/**").hasAuthority("ADMIN")
-                        .requestMatchers("/index", "/login", "/register").permitAll()
+                        .requestMatchers("/**").permitAll()
                         .anyRequest().authenticated()
                 )
                 .formLogin((formLogin) ->
                         formLogin
                                 .loginPage("/login")
                                 .failureUrl("/login?error=true")
-                                .defaultSuccessUrl("/index", true)
+                                .defaultSuccessUrl("/", true)
                                 .usernameParameter("userId")
                                 .passwordParameter("userPw")
                                 .permitAll()
@@ -53,7 +48,6 @@ public class CustomSecurityConfig {
                         .logoutSuccessUrl("/login")
                         .permitAll()
                 );
-
 
         return http.build();
     }
