+++ src/main/java/com/myapp/team/user/mapper/AdminMapper.java
@@ -1,0 +1,9 @@
+package com.myapp.team.user.mapper;
+
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+@Mapper
+public interface AdminMapper {
+    void updateUserRole(@Param("userId") String userId, @Param("role") String role);
+}
