+++ src/main/resources/application.properties
@@ -13,3 +13,4 @@ logging.level.com.myapp.team.user.mapper=TRACE
 spring.thymeleaf.prefix=classpath:/templates/
 spring.thymeleaf.suffix=.html
 spring.thymeleaf.cache=false
+
