+++ src/main/resources/templates/index.html
@@ -61,7 +61,7 @@
     <a href="#">Products</a>
     <a href="#">About Us</a>
     <a href="#">Contact</a>
-    <span sec:authentication="principal.authorities"></span>
+<!--    <span sec:authentication="principal.authorities"></span>-->
 </nav>
 <section>
     <div class="product">
