+++ src/main/resources/templates/login.html
@@ -54,7 +54,7 @@
     <input type="password" id="userPw" name="userPw" required>
     <input type="submit" value="Login">
     <input type="button" value="회원가입" onclick="goToRegister()">
-
 </form>
+
 </body>
 </html>
